@@ -26,7 +26,7 @@ setup(
     include_package_data=True,
     install_requires=install_requires,
     license="Apache License 2.0",
-    python_requires=">=3.6",
+    python_requires=">=3.7",
     zip_safe=False,
     keywords=["appdaemon", "home", "automation"],
     entry_points={"console_scripts": ["appdaemon = appdaemon.__main__:main"]},
@@ -35,9 +35,9 @@ setup(
         "Intended Audience :: Developers",
         "License :: OSI Approved :: Apache Software License",
         "Natural Language :: English",
-        "Programming Language :: Python :: 3.6",
         "Programming Language :: Python :: 3.7",
         "Programming Language :: Python :: 3.8",
+        "Programming Language :: Python :: 3.0",
         "Topic :: Home Automation",
     ],
 )
