@@ -14,6 +14,10 @@ function getCookie(cname) {
     return "";
 }
 
+function setCookie(cname, cvalue) {
+    document.cookie = cname + "=" + cvalue + "; expires=Sat, 1 Jan 2050 12:00:00 UTC;";
+}
+
 function get_monitored_entities(widgets)
 {
     index = 0;
@@ -49,6 +53,9 @@ var DashStream = function(transport, protocol, domain, port, title, widgets)
 
         // subscribe to all events
 
+        self.stream.listen_event('*', 'ad_dashboard', self.update_dash);
+
+        //deprecated. Will be removed.
         self.stream.listen_event('*', '__HADASHBOARD_EVENT', self.update_dash);
 
         // Subscribe to just the entities we care about for this dashboard
@@ -88,7 +95,11 @@ var DashStream = function(transport, protocol, domain, port, title, widgets)
     this.update_dash = function(msg)
     {
         data = msg.data;
-        if (data.event_type === "__HADASHBOARD_EVENT")
+        if ( (data.event_type === "__HADASHBOARD_EVENT" ||
+              data.event_type === "ad_dashboard")  &&
+             ((data.data.deviceid && data.data.deviceid === myDeviceID) ||
+              (data.data.dashid && title.includes(data.data.dashid)) ||
+              (!data.data.deviceid && !data.data.dashid)))
         {
             if (data.data.command === "navigate")
             {
@@ -399,13 +410,26 @@ var WidgetBase = function(widget_id, url, skin, parameters, monitored_entities, 
     {
         if ("selector" in callbacks[i])
         {
-            $(callbacks[i].selector).on(callbacks[i].action, (
-                function (callback, ch, params) {
-                    return function () {
-                        callback(ch, params)
-                    };
-                }(callbacks[i].callback, child, callbacks[i].parameters))
-            );
+            if ("DOMEventData" in callbacks[i] && callbacks[i].DOMEventData === true)
+            {
+
+                var data = {};
+                data.parameters = callbacks[i].parameters;
+
+                $(callbacks[i].selector).on(callbacks[i].action, data, callbacks[i].callback);
+
+            }
+
+            else
+            {
+                $(callbacks[i].selector).on(callbacks[i].action, (
+                    function (callback, ch, params) {
+                        return function () {
+                            callback(ch, params)
+                        };
+                    }(callbacks[i].callback, child, callbacks[i].parameters))
+                );
+            }
         }
         else if ("observable" in callbacks[i])
         {
