@@ -619,6 +619,18 @@ class Dashboard:
             dash["head_includes"] = css_vars["head_includes"]
         else:
             dash["head_includes"] = []
+
+        #
+        # adds custom_javascripts to the head includes if they exist
+        #
+        custom_js = os.path.join(self.config_dir, "custom_javascript")
+        if os.path.isdir(custom_js):
+            for filename in os.listdir(custom_js):
+                if filename.endswith(".js"):
+                    dash["head_includes"].append(
+                        '<script type="text/javascript" src="custom_javascript/{}"></script>'.format(filename)
+                    )
+
         if "body_includes" in css_vars and css_vars["body_includes"] is not None:
             dash["body_includes"] = css_vars["body_includes"]
         else:
@@ -730,7 +742,7 @@ class Dashboard:
                         widgets[widget] = {"js": js, "css": css, "html": html}
         return widgets
 
-    def _list_dashes(self):
+    def list_dashes(self):
         if not os.path.isdir(self.dashboard_dir):
             return {}
 
@@ -816,7 +828,7 @@ class Dashboard:
 
         dash = self._get_dash(name, skin, skindir)
         if dash is None:
-            dash_list = self._list_dashes()
+            dash_list = self.list_dashes()
             return {"errors": ["Dashboard has errors or is not found - check log for details"], "dash_list": dash_list}
 
         params = dash
@@ -945,7 +957,7 @@ class Dashboard:
     def get_dashboard_list(self, paramOverwrite=None):
 
         if paramOverwrite is None:
-            dash = self._list_dashes()
+            dash = self.list_dashes()
         else:
             dash = paramOverwrite
 
