@@ -415,7 +415,7 @@ class Logging:
 
         return logger
 
-    async def get_admin_logs(self, maxlines=50):
+    async def get_admin_logs(self, maxlines=100):
         return await utils.run_in_executor(self, self._get_admin_logs, maxlines)
 
     def _get_admin_logs(self, maxlines):
@@ -425,16 +425,22 @@ class Logging:
             logs[log] = {}
             logs[log]["name"] = self.config[log]["name"]
             logs[log]["lines"] = self.read_logfile(log)
+            while len(logs[log]["lines"]) > maxlines:
+                logs[log]["lines"].pop(0)
+
         for log in self.config:
             if log not in logs:
                 logs[log] = {}
                 logs[log]["name"] = self.config[log]["name"]
                 logs[log]["lines"] = self.read_logfile(log)
+                while len(logs[log]["lines"]) > maxlines:
+                    logs[log]["lines"].pop(0)
+
         return logs
 
     def read_logfile(self, log):
         if self.is_alias(log):
-            return None
+            return []
         if self.config[log]["filename"] == "STDOUT" or self.config[log]["filename"] == "STDERR":
             return []
         else:
@@ -595,6 +601,8 @@ class Logging:
 
         """
 
+        executed = False
+
         if not isinstance(handles, list):
             handles = [handles]
 
@@ -603,5 +611,13 @@ class Logging:
                 if name in self.AD.callbacks.callbacks and handle in self.AD.callbacks.callbacks[name]:
                     del self.AD.callbacks.callbacks[name][handle]
                     await self.AD.state.remove_entity("admin", "log_callback.{}".format(handle))
+                    executed = True
                 if name in self.AD.callbacks.callbacks and self.AD.callbacks.callbacks[name] == {}:
                     del self.AD.callbacks.callbacks[name]
+
+        if not executed:
+            self.logger.warning(
+                "Invalid callback handles '{}' in cancel_log_callback() from app {}".format(handles, name)
+            )
+
+        return executed
