@@ -2,7 +2,6 @@ import traceback
 import datetime
 from datetime import timedelta
 import pytz
-import astral
 import random
 import uuid
 import re
@@ -12,6 +11,7 @@ from collections import OrderedDict
 
 import appdaemon.utils as utils
 from appdaemon.appdaemon import AppDaemon
+from astral.location import Location, LocationInfo
 
 
 class Scheduler:
@@ -92,12 +92,29 @@ class Scheduler:
         self.stopping = True
 
     async def cancel_timer(self, name, handle):
+        executed = False
         self.logger.debug("Canceling timer for %s", name)
-        if name in self.schedule and handle in self.schedule[name]:
+        if self.timer_running(name, handle):
             del self.schedule[name][handle]
             await self.AD.state.remove_entity("admin", "scheduler_callback.{}".format(handle))
+            executed = True
+
         if name in self.schedule and self.schedule[name] == {}:
             del self.schedule[name]
+
+        if not executed:
+            self.logger.warning("Invalid callback handle '{}' in cancel_timer() from app {}".format(handle, name))
+
+        return executed
+
+    def timer_running(self, name, handle):
+        """Check if the handler is valid
+        by ensuring the timer is still running"""
+
+        if name in self.schedule and handle in self.schedule[name]:
+            return True
+
+        return False
 
     # noinspection PyBroadException
     async def exec_schedule(self, name, args, uuid_):
@@ -107,6 +124,11 @@ class Scheduler:
                 #
                 # it's a "duration" entry
                 #
+
+                # first remove the duration parameter
+                if args["kwargs"].get("__duration"):
+                    del args["kwargs"]["__duration"]
+
                 executed = await self.AD.threading.dispatch_worker(
                     name,
                     {
@@ -130,7 +152,9 @@ class Scheduler:
                     if remove is True:
                         await self.AD.state.cancel_state_callback(args["kwargs"]["__handle"], name)
 
-                        if "__timeout" in args["kwargs"]:  # meaning there is a timeout for this callback
+                        if "__timeout" in args["kwargs"] and self.timer_running(
+                            name, args["kwargs"]["__timeout"]
+                        ):  # meaning there is a timeout for this callback
                             await self.cancel_timer(name, args["kwargs"]["__timeout"])  # cancel it as no more needed
 
             elif "__state_handle" in args["kwargs"]:
@@ -215,56 +239,51 @@ class Scheduler:
         if longitude < -180 or longitude > 180:
             raise ValueError("Longitude needs to be -180 .. 180")
 
-        elevation = self.AD.elevation
+        self.location = Location(LocationInfo("", "", self.AD.tz.zone, latitude, longitude))
 
-        self.location = astral.Location(("", "", latitude, longitude, self.AD.tz.zone, elevation))
+    def sun(self, type: str, secs_offset: int):
+        return self.get_next_sun_event(type, secs_offset) + datetime.timedelta(seconds=secs_offset)
 
-    def sun(self, type, offset):
-        if offset < 0:
-            # For negative offset we need to look forward to the next event after the current one
-            return self.get_next_sun_event(type, 1) + datetime.timedelta(seconds=offset)
-        else:
-            # Positive or zero offset so no need to specify anything special
-            return self.get_next_sun_event(type, 0) + datetime.timedelta(seconds=offset)
-
-    def get_next_sun_event(self, type, offset):
+    def get_next_sun_event(self, type: str, day_offset: int):
         if type == "next_rising":
-            return self.next_sunrise(offset)
+            return self.next_sunrise(day_offset)
         else:
-            return self.next_sunset(offset)
+            return self.next_sunset(day_offset)
 
-    def next_sunrise(self, offset=0):
-        mod = offset
+    def next_sunrise(self, offset: int = 0):
+        day_offset = 0
         while True:
             try:
+                candidate_date = (self.now + datetime.timedelta(days=day_offset)).astimezone(self.AD.tz).date()
                 next_rising_dt = self.location.sunrise(
-                    (self.now + datetime.timedelta(seconds=offset) + datetime.timedelta(days=mod)).date(), local=False,
+                    date=candidate_date, local=False, observer_elevation=self.AD.elevation
                 )
-                if next_rising_dt > self.now:
+                if next_rising_dt + datetime.timedelta(seconds=offset) > (self.now + datetime.timedelta(seconds=1)):
                     break
-            except astral.AstralError:
+            except ValueError:
                 pass
-            mod += 1
+            day_offset += 1
 
         return next_rising_dt
 
-    def next_sunset(self, offset=0):
-        mod = offset
+    def next_sunset(self, offset: int = 0):
+        day_offset = 0
         while True:
             try:
+                candidate_date = (self.now + datetime.timedelta(days=day_offset)).astimezone(self.AD.tz).date()
                 next_setting_dt = self.location.sunset(
-                    (self.now + datetime.timedelta(seconds=offset) + datetime.timedelta(days=mod)).date(), local=False,
+                    date=candidate_date, local=False, observer_elevation=self.AD.elevation
                 )
-                if next_setting_dt > self.now:
+                if next_setting_dt + datetime.timedelta(seconds=offset) > (self.now + datetime.timedelta(seconds=1)):
                     break
-            except astral.AstralError:
+            except ValueError:
                 pass
-            mod += 1
+            day_offset += 1
 
         return next_setting_dt
 
     @staticmethod
-    def get_offset(kwargs):
+    def get_offset(kwargs: dict):
         if "offset" in kwargs["kwargs"]:
             if "random_start" in kwargs["kwargs"] or "random_end" in kwargs["kwargs"]:
                 raise ValueError(
@@ -520,7 +539,7 @@ class Scheduler:
                     # Nothing to do, lets wait for a while, we will get woken up if anything new comes along
                     delay = idle_time
 
-                # Initially we don't want to skip over any events that haven;t had a chance to be registered yet, but now
+                # Initially we don't want to skip over any events that haven't had a chance to be registered yet, but now
                 # we can loosen up a little
                 idle_time = 60
 
@@ -590,7 +609,7 @@ class Scheduler:
         return self.next_sunrise() < self.next_sunset()
 
     async def info_timer(self, handle, name):
-        if name in self.schedule and handle in self.schedule[name]:
+        if self.timer_running(name, handle):
             callback = self.schedule[name][handle]
             return (
                 self.make_naive(callback["timestamp"]),
