@@ -653,7 +653,10 @@ class Threading:
                     # Or we won't if they are not.
                     # Either way, we cancel the old timer
                     #
-                    await self.AD.sched.cancel_timer(name, kwargs["__duration"])
+                    if self.AD.sched.timer_running(name, kwargs["__duration"]):
+                        await self.AD.sched.cancel_timer(name, kwargs["__duration"])
+
+                    del kwargs["__duration"]
 
                 #
                 # Check if we care about the change
