@@ -63,6 +63,7 @@ function basemedia(widget_id, url, skin, parameters)
     {
         self.entity = state.entity_id;
         self.level = state.attributes.volume_level;
+        self.state = state;
         set_view(self, state)
         if ("dump_capabilities" in self.parameters && self.parameters["dump_capabilities"] == "1")
         {
@@ -77,6 +78,7 @@ function basemedia(widget_id, url, skin, parameters)
     function OnStateUpdate(self, state)
     {
         self.level = state.attributes.volume_level;
+        self.state = state;
         set_view(self, state)
     }
 
@@ -99,11 +101,6 @@ function basemedia(widget_id, url, skin, parameters)
             if (is_supported(self, "PAUSE"))
             {
                 args = self.parameters.post_service_pause;
-                self.call_service(self, args)
-            }
-            else if (is_supported(self, "STOP"))
-            {
-                args = self.parameters.post_service_stop;
                 self.call_service(self, args)
             }
             else if (is_supported(self, "STOP"))
@@ -243,7 +240,7 @@ function basemedia(widget_id, url, skin, parameters)
                 "SHUFFLE_SET": 32768
             };
 
-        var supported = self.entity_state[parameters.entity].attributes.supported_features;
+        var supported = self.state.attributes.supported_features;
 
         if (attr in support)
         {
@@ -267,7 +264,7 @@ function basemedia(widget_id, url, skin, parameters)
     function display_supported_functions(self)
     {
         console.log(self.parameters.entity);
-        console.log("Supported Features: " + self.entity_state[parameters.entity].attributes.supported_features);
+        console.log("Supported Features: " + self.state.attributes.supported_features);
         console.log("PAUSE: " + is_supported(self, "PAUSE"))
         console.log("SEEK: " + is_supported(self, "SEEK"))
         console.log("VOLUME_SET: " + is_supported(self, "VOLUME_SET"))
