@@ -58,7 +58,7 @@ master_doc = "index"
 
 # General information about the project.
 project = "AppDaemon"
-copyright = "2019, Andrew Cockburn"
+copyright = "2021, Andrew Cockburn"
 author = "Andrew Cockburn"
 
 # The version info for the project you're documenting, acts as replacement for
@@ -66,9 +66,9 @@ author = "Andrew Cockburn"
 # built documents.
 #
 # The short X.Y version.
-version = "4.0.6"
+version = "4.1.0"
 # The full version, including alpha/beta/rc tags.
-release = "4.0.6"
+release = "4.1.0"
 
 # The language for content autogenerated by Sphinx. Refer to documentation
 # for a list of supported languages.
